flight-service/src/main/java/com/wipro/flightms/controller/LocationController.java
package com.wipro.flightms.controller;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.wipro.flightms.entity.Location;
import com.wipro.flightms.service.LocationService;

@RestController
@RequestMapping("/locations")
public class LocationController {
	@Autowired
    private LocationService locationService;

    @GetMapping
    public List<Location> getAllLocations() {
        return locationService.getAllLocations();
    }

}
